uploaded sum.c
The first leet code problem i solved using the knowledge and experience i gained from my mentor in understanding the C programming language. It takes the size, elements of an array and then ask the user the *target* value which is an number and the program finds the index of those elements which sums up the target value.